requirements.txt
livekit>=1.0
livekit-agents~=1.0
livekit-plugins-google~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0

